requirements.txt
beautifulsoup4==4.12.2
lxml==4.9.3
fastapi==0.104.1
uvicorn==0.24.0.post1
langchain==0.0.331
nltk==3.8.1
pandas==2.1.2
playwright==1.39.0
pyspellchecker==0.7.2
selenium==4.15.2
textblob==0.17.1
webdriver-manager==4.0.1
openai==0.28.1
opencv-python==4.8.1.78
Pillow==10.1.0
moviepy==1.0.3
tiktoken==0.5.1
pdf2image==1.16.3
fuzzywuzzy==0.18.0

